+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2
+- Adding top photo for Venue
+- Adding Category object
+- Minor updates to links in README
+
 ## 0.0.1
 - Initial version
 - Support for users/X, venues/search